requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-appconf==1.0.3
django-debug-toolbar==1.11
django-imagekit==2.0.4
django-taggit==1.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
oauthlib==3.1.0
pilkit==2.0
Pillow==6.0.0
PyJWT==1.7.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
redis==3.1.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
